+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fix: keyring handling in `icloud`. Rel to [#871](https://github.com/icloud-photos-downloader/icloud_photos_downloader/issues/871)
+
 ## 1.20.0 (2024-06-16)
 
 - feature: customize choice and the order of checking for password with `--password-provider` parameter